+++ lib/async/container/threaded.rb
@@ -211,11 +211,6 @@ module Async
 					@thread.kill
 				end
 				
-				# Raise {Restart} in the child thread.
-				def restart!
-					@thread.raise(Restart)
-				end
-				
 				# Wait for the thread to exit and return he exit status.
 				# @asynchronous This method may block.
 				#